Extract <Table> from App.tsx
